requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2026.2.25
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
deprecation==2.1.0
Flask==3.0.3
fonttools==4.61.1
grpcio==1.78.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.9.2
mplfinance==0.12.10b0
numpy==2.4.2
packaging==26.0
pandas==2.2.2
pillow==12.1.1
protobuf==4.25.8
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
sentry-sdk==2.53.0
six==1.17.0
t-tech-investments==0.3.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.6
numpy
scikit-learn
